learn how to add double array

+++ searching/level 1/arr.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int arr[5];
+    double arr[5];
     cin >> arr[0] >> arr[1] >> arr[2] >> arr[3] >> arr[4];
